break up homepage into loadable components
